test: 신고 조회 테스트 코드 작성
